heronPipeline/src/functions/checkLqpMetaDataIsPresent/requirements.txt
pandas
numpy
boto